Move Print connect to kapture-backend
