+++ ui/src/components/index.js
@@ -2,3 +2,4 @@ export { Header } from './app/Header';
 export { ActivitySearch } from './activity/ActivitySearch';
 export { ActivityHeader } from './activity/ActivityHeader';
 export { PlateSizeForm } from './activity/PlateSizeForm';
+export { Print } from './print';
