+++ ui/src/implementations/kapture-backend/components/index.js
@@ -3,3 +3,4 @@ export { ActivitySearch } from './ActivitySearchConnected';
 export { ActivityHeader } from './ActivityHeaderConnected';
 export { ActivityDetails } from './ActivityDetails';
 export { PlateSizeForm } from '../../../components';
+export { Print } from './PrintConnected';
